updated pledge forms and project details

+++ src/pages/ProjectPage.jsx
@@ -1,12 +1,15 @@
 import React, { useState, useEffect } from "react";
 import { useParams, useHistory, Link } from "react-router-dom";
 import PledgeForm from "../components/PledgeForm/PledgeForm";
+// import Logo from "./components/Logo/Logo";
 // import image from "../image.png";
-// import Logo from "./public/Logo.png";
 // import { oneProject } from "../data";
 
 function ProjectPage() {
-  const [projectData, setProjectData] = useState({ pledges: [] });
+  const [projectData, setProjectData] = useState({
+    loading: true,
+    pledges: [],
+  });
   const { id } = useParams();
 
   const token = window.localStorage.getItem("token");
@@ -36,31 +39,37 @@ function ProjectPage() {
     history.push("/");
   };
 
+  if (projectData.loading) {
+    return "Loading ...";
+  }
+
   return (
     <div>
+      <img src={projectData.image} />
       <h2>{projectData.title}</h2>
       <h3>
         Created at:{" "}
         {projectData.date_created ? projectData.date_created.substr(0, 10) : ""}
       </h3>
       <h3>{`Status: ${projectData.is_open}`}</h3>
-      {/* {<h3>Pledges:</h3>
-        <ul>
+      {projectData.pledges.length > 0 ? <h3>Pledges:</h3> : null}
+      <ul>
         {projectData.pledges.map((pledgeData, key) => {
-            return (
+          return (
             <li>
-            ${pledgeData.amount} from {pledgeData.supporter}
+              ${pledgeData.amount} from {pledgeData.supporter}
+              <br />
+              {pledgeData.comment}
             </li>
-            );   
+          );
         })}
-        </ul>} */}
-      {/* <button type="edit" onClick={handleSubmit}>
-        Edit{" "}
-      </button> */}
+      </ul>
 
       <PledgeForm projectId={id} />
       <hr />
-      <Link to={`/editproject/${id}`}>Edit</Link>
+      <button>
+        <Link to={`/editproject/${id}`}>Edit</Link>
+      </button>
       <button type="delete" onClick={deleteData}>
         Delete
       </button>
